requirements.txt
langchain
langchain-community
llama-parse 
qdrant_client
python-dotenv
langchain-groq
chainlit
fastembed
unstructured[md]
langchain-openai

